+++ xcube/cli/gen2.py
@@ -167,9 +167,7 @@ def gen2(request_path: str,
                              'Cube generation failed.'
                              if error is None else f'{error}')
         click_error = click.ClickException(message)
-        if error is not None:
-            raise click_error from error
-        raise click_error
+        raise click_error from error
 
 
 if __name__ == '__main__':